refactored: working
